requirements.txt
fastapi==0.115.5
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
uvicorn
Pillow==11.0.0
google-generativeai==0.8.3
python-multipart